review updates 2

+++ client/src/com/aerospike/client/exp/Expression.java
@@ -96,7 +96,7 @@ public final class Expression implements Serializable {
 	 * Write expression in wire protocol.
 	 * For internal use only.
 	 */
-	public long write(Command cmd) {
+	public int write(Command cmd) {
 		cmd.writeExpHeader(bytes.length);
 		System.arraycopy(bytes, 0, cmd.dataBuffer, cmd.dataOffset, bytes.length);
 		cmd.dataOffset += bytes.length;
